Updates to eligibility criteria
Main updates:
Added two screens for the A/B test for the 70% child-facing question
A is testing a declaration pattern
B is testing a question pattern (Q: Would we need an "I don't know" option here?)
Added the new question flow for the 6 month question
Test the flow here
Q: On the second screen of this flow (contract types), would we need an "other" choice? Where would we take the user here?
Two new exit points
If a user selects "I don't know", they'll be taken to a screen where we'll explain that they'll need to confirm this information to continue.
If a user selects "I cannot confirm this information" on either the A or B screen of the 70% child-facing question, they'll be taken to an ineligible screen.
If a user selects that they've worked in a permanent role (that was 70% child-facing) in the past 6 months, they'll be taken to an ineligible screen.

Smaller updates:
Added the confirmations to the check your answers page
Added a new "documents" section to the landing page to display the consent form and the privacy notice